[cilium] Support Gateway API
## Context
The high-level intent behind [this](https://github.com/hcloud-talos/terraform-hcloud-talos/issues/278) issue was to make it easy for a user of this library to use Cilium as Gateway API implementation. The suggestion was to allow customising cilium configuration without having to add a lot of boilerplate.

## Change
On further experimenting, this approach wouldn't scale well as:
1. Enabling gateway API involves setting a collection of flags and installing dependencies
2. If cilium was installed (with gateway enabled) before dependencies are installed, gateways wouldn't work even if the dependencies were installed later. This is because the Cilium operator doesn't continuously reconcile.

Instead, this diff abstracts away gateway API configuration with a variable and internally installs required dependencies.

NOTE: We install the experimental yaml and not the main yaml as the former includes tlsroute CRDs which are required for the Gateway API regardless of whether it uses TLS or not.

## Test Plan

```
terraform validate
```

\+
Use it on my cluster and validate that gateway is correctly configured.